add auth routines

+++ routines_ms/mutations.js
@@ -1,7 +1,9 @@
 const axios = require('axios');
 //aqui toca reemplazar ${process.env.ROUTINES_URL}
 const exerciseMutations = {
-  createExercise: async (_, { exerciseName, muscularGroup }) => {
+  createExercise: async (_, { exerciseName, muscularGroup }, { user, role }) => {
+    if (!user) throw new Error("No autenticado");
+    if (role !== "admin" && role !== "coach") throw new Error("No tienes permisos");
     const payload = {
       exercise_name: exerciseName,
       muscular_group: muscularGroup.map(mg => ({
@@ -35,7 +37,9 @@ const exerciseMutations = {
     }
   },
   
-  updateExercise: async (_, { id, exerciseName, muscularGroup }) => {
+  updateExercise: async (_, { id, exerciseName, muscularGroup }, { user, role }) => {
+    if (!user) throw new Error("No autenticado");
+    if (role !== "admin" && role !== "coach") throw new Error("No tienes permisos");
     const payload = {
       id: id,
       exercise_name: exerciseName,
@@ -67,7 +71,9 @@ const exerciseMutations = {
     }
   },
   
-  deleteExercise: async (_, { id }) => {
+  deleteExercise: async (_, { id }, { user, role }) => {
+    if (!user) throw new Error("No autenticado");
+    if (role !== "admin" && role !== "coach") throw new Error("No tienes permisos");
     try {
       await axios.delete(`${process.env.ROUTINES_URL}/exercises/${id}`);
       return id;
@@ -79,7 +85,9 @@ const exerciseMutations = {
 };
 
 const routineMutations = {
-  createRoutine: async (_, { routineName, routineDifficulty, routineExercises }) => {
+  createRoutine: async (_, { routineName, routineDifficulty, routineExercises }, { user, role }) => {
+    if (!user) throw new Error("No autenticado");
+    if (role !== "admin" && role !== "coach") throw new Error("No tienes permisos");
     const payload = {
       routine_name: routineName,
       routine_difficulty: routineDifficulty,
@@ -112,7 +120,9 @@ const routineMutations = {
     }
   },
 
-  updateRoutine: async (_, { id, routineName, routineDifficulty, routineExercises }) => {
+  updateRoutine: async (_, { id, routineName, routineDifficulty, routineExercises }, { user, role }) => {
+    if (!user) throw new Error("No autenticado");
+    if (role !== "admin" && role !== "coach") throw new Error("No tienes permisos");
     const payload = {
       routine_name: routineName,
       routine_difficulty: routineDifficulty,
@@ -139,7 +149,9 @@ const routineMutations = {
     }
   },
   
-  deleteRoutine: async (_, { id }) => {
+  deleteRoutine: async (_, { id }, { user, role }) => {
+    if (!user) throw new Error("No autenticado");
+    if (role !== "admin" && role !== "coach") throw new Error("No tienes permisos");
     try {
       await axios.delete(`${process.env.ROUTINES_URL}/routines/${id}`);
       return id;
